fixed error when deleting categories that have specific interests inside

+++ includes/deleteInterest.php
@@ -4,9 +4,13 @@ require_once 'dbhInc.php';
 
 if(isset($_GET['interestsId'])){
     $interestsId = $_GET['interestsId'];
-    $sql = "DELETE FROM interests WHERE interestsId = '$interestsId';";
-    $result = mysqli_query($conn, $sql);//fatal error here when deleting the first item in the interest category list, all others delete fine
+    $sql = "DELETE FROM items WHERE interestsId = '$interestsId';";
+    $result = mysqli_query($conn, $sql);
 
+    if ($result) {
+        $sql = "DELETE FROM interests WHERE interestsId = '$interestsId';";
+        $result = mysqli_query($conn, $sql);
+    
     if ($result) {
         mysqli_close($conn);
         header("location: ../userProfile.php");
@@ -14,4 +18,5 @@ if(isset($_GET['interestsId'])){
     }else {
         echo "Error deleting record";
     }
+}
 }
